Corrected all the imports to the refactored componenets
