+++ src/components/actors/actorCard/index.jsx
@@ -11,7 +11,7 @@ import PlaylistIcon from "@mui/icons-material/PlaylistAdd";
 import CalendarIcon from "@mui/icons-material/CalendarTodayTwoTone";
 import StarRateIcon from "@mui/icons-material/StarRate";
 import Grid from "@mui/material/Grid";
-import img from '../../images/film-poster-placeholder.png'
+import img from '../../../images/film-poster-placeholder.png'
 import { Link } from "react-router-dom";
 import BadgeIcon from '@mui/icons-material/Badge';
