+++ src/components/movies/filterMoviesCard/index.jsx
@@ -7,9 +7,9 @@ import MenuItem from "@mui/material/MenuItem";
 import TextField from "@mui/material/TextField";
 import FilterAltIcon from '@mui/icons-material/FilterAlt';
 import SortIcon from '@mui/icons-material/Sort';
-import { getGenres } from "../../api/tmdb-api";
+import { getGenres } from "../../../api/tmdb-api";
 import { useQuery } from "react-query";
-import Spinner from '../spinner'
+import Spinner from '../../spinner'
 import FormControl from "@mui/material/FormControl";
 import Select from "@mui/material/Select";
