+++ src/components/movies/movieFilterUI/index.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import FilterCard from "../../filterMoviesCard";
+import FilterCard from "../../movies/filterMoviesCard";
 import Fab from "@mui/material/Fab";
 import Drawer from "@mui/material/Drawer";
