+++ src/pages/movieDetailsPage.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useParams } from "react-router-dom";
 import MovieDetails from "../components/movies/movieDetails";
-import PageTemplate from "../components/templateMoviePage";
+import PageTemplate from "../components/movies/templateMoviePage";
 import { getMovie } from '../api/tmdb-api'
 import { useQuery } from "react-query";
 import Spinner from '../components/spinner'
