+++ src/pages/movieReviewPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useLocation } from "react-router-dom";
-import PageTemplate from "../components/templateMoviePage";
+import PageTemplate from "../components/movies/templateMoviePage";
 import MovieReview from "../components/movies/movieReview";
 
 const MovieReviewPage = (props) => {
